Quick and dirty fix for initializing configspec.ini
When I downloaded the compiled gooeypi.exe binary, I found it can not start up. After I studied the source code, I realized it requires an configspec.ini, which does not come with the binary. It's better for the program to initialize the file automatically, so some code to create the spec is added here.